Move client msg. processing and server resp. formation to 'game'
Currently, everything related to processing the client message/
request and forming the server message/response to be sent back to
the client is done via the Telnet handler Greenlet object. This
shouldn't be the case, it should be handled by the game logic itself.

Move the request handling and response creation to the 'GameSession'
class in the new 'game' module. This is the first step towards
expanding the project into the actual game.