Implement user action via 'GameSession' object
Rather than asking the 'GameSession' object for the response based
on the input to 'apply', give the 'GameSession' object the socket
so that it can reference it directly. Simply execute the 'GameSession'
object's run from 'handle', and let the 'GameSession' decide
everything from there.

Additionally, wrap the socket via a 'GameSocket' object, so we can
add our own logic on top of it. Might be preferable to just have
'GameSocket' be a subclass of the socket's class, though I don't
think we can control the type of the object that is returned from
'accept'... Likely need to continue wrapping.