Initial implementation of authentication system
When creating user, pass base64-encoded key (base64-encoded to
avoid newlines, so no need to call 'recv' multiple times).
    * Also made 'recv' calls attempt to get a max of 10000
      chars., for very large keys - though may be unnecessary...
        *** Look into ideal maximum upper bound, 'select', etc.
    * The decoded version is stored in the database.

When you type in your name next time, it grabs the public key,
encrypts a temporary secret, and then requests that you respond
with the decrypted version of the secret. Note that the encrypted
secret is base64-encoded, to avoid console rendering and/or
copy-paste issues with the binary data.

The command to perform the above should be as follows...
echo '<FROM SERVER>' | base64 -d > challenge && openssl rsautl
-decrypt -in challenge -inkey mykey.pem -out challenge-dec &&
cat challenge-dec

* This implementation works, but needs cleanup...