Fix gevent startup issues
Gevent tends to not be very happy if the monkey patching isn't
done at a certain time. The import order changes done for the
Pylint fixes got Gevent into a bad state, where the interpreter
just dies.

Move the monkey patching to the top of the entrypoint script,
which will also keep Pylint happy.