Use Singleton pattern for Config
There's no reason we need more than one Config object, and as we
start to add more data, it could become costly to load it several
times. If we need to reference it from several files, it seems
better to just reference the Singleton Config as opposed to
reinstantiating/reloading it over and over again.